Dokončený filter, rozpracovaný login

+++ backend/parking-app/src/main/java/com/example/parkingapp/reservation/ReservationService.java
@@ -4,5 +4,5 @@ import com.example.parkingapp.parkingPlace.ParkingPlace;
 
 public interface ReservationService {
     void saveReservation(Reservation reservation);
-    Reservation updateReservation(int reservationId, Reservation reservation);
+    Reservation updateReservation(long reservationId, Reservation reservation);
 }
